大部分功能实现

+++ app/src/main/java/com/maureen/wandevelop/entity/MessageInfo.kt
@@ -12,14 +12,14 @@ data class MessageInfo(
     val category: Int,
     val date: Long,
     val fromUser: String,
-    val fromUserId: Int,
+    val fromUserId: Long,
     val fullLink: String,
-    val id: Int,
+    val id: Long,
     val isRead: Int,
     val link: String,
     val message: String,
     val niceDate: String,
     val tag: String,
     val title: String,
-    val userId: Int
+    val userId: Long
 )
